fix(investments): extract rangeToDates to lib/utils — server actions must be async
