+++ frontend/lib/actions/investments.ts
@@ -1,15 +1,9 @@
 "use server";
 
-export type Range = "1W" | "1M" | "3M" | "1Y" | "ALL";
+import { rangeToDates, type Range } from "@/lib/utils/date-ranges";
 
-export function rangeToDates(range: Range, now: Date = new Date()) {
-  const to = now.toISOString().slice(0, 10);
-  if (range === "ALL") return { from: "2010-01-01", to };
-  const days = { "1W": 7, "1M": 30, "3M": 90, "1Y": 365 }[range];
-  const d = new Date(now);
-  d.setUTCDate(d.getUTCDate() - days);
-  return { from: d.toISOString().slice(0, 10), to };
-}
+// Re-export for consumers that import Range from this module
+export type { Range } from "@/lib/utils/date-ranges";
 
 export async function fetchHistoryRange(range: Range) {
   const { getPortfolioHistory } = await import("@/lib/api/investments");
